Add draftjs utils to libs, restructure libs
