+++ src/lib/utils.js
@@ -10,16 +10,7 @@ const logDebug = (fn) => {
   }
 }
 
-const ensureLoaders = (loaders) => {
-  if (!loaders) {
-    throw new Error('LOADERS_NOT_INITIALIZED')
-  }
-
-  return loaders
-}
-
 module.exports = {
   exceptTest,
   logDebug,
-  ensureLoaders,
 }
